projects/tricket/tricket-tests/requirements.txt
supabase
rich
typer
python-dotenv
